+++ modules/Accounting/Incomes.php
@@ -179,6 +179,8 @@ if ( ! $_REQUEST['modfunc'] )
 		echo '</form>';
 	}
 
+	$options = [];
+
 	if ( ! $_REQUEST['print_statements'] && AllowEdit() )
 	{
 		echo '<form action="' . PreparePHP_SELF() . '" method="POST">';
@@ -186,10 +188,6 @@ if ( ! $_REQUEST['modfunc'] )
 		DrawHeader( '', SubmitButton() );
 
 		$options = [ 'valign-middle' => true ];
-	}
-	else
-	{
-		$options = [ 'center' => false, 'add' => false ];
 	}
 
 	ListOutput( $RET, $columns, 'Income', 'Incomes', $link, [], $options );