+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 11.7
 - CSS FlatSIS theme: add box-shadow to colorBox in colors.css, stylesheet.css & zresponsive.css
 - Allow display of `$link['add']` (or remove) on PDF or if not allowed to edit in ListOutput.fnc.php
 - Force display of `$link['add']` on PDF or if not allowed to edit in DailyTransactions.php, Percent.php, InputFinalGrades.php & ProgressReports.php
+- Remove useless center option in ListOutput() & program wide
 
 Changes in 11.6
 ---------------